Move command to separate directory
Also, add command to pyproject.toml to make readily accessible.
