+++ magma_ff/create_metawfr.py
@@ -6,7 +6,6 @@
 #
 ################################################
 
-import argparse
 import datetime
 import json
 import uuid
@@ -16,7 +15,7 @@ from dcicutils import ff_utils
 
 from magma_ff.metawfl import MetaWorkflow
 from magma_ff.metawflrun import MetaWorkflowRun
-from magma_ff.utils import JsonObject, keep_last_item, get_auth_key, make_embed_request
+from magma_ff.utils import JsonObject, keep_last_item, make_embed_request
 
 
 FILE_FORMAT = "file_format"
@@ -1250,39 +1249,3 @@ class InputPropertiesFromSample:
     def input_sample_uuids(self):
         """Sample UUID"""
         return [self.sample[self.UUID]]
-
-
-def main() -> None:
-    parser = argparse.ArgumentParser(description="Create MetaWorkflowRun")
-    parser.add_argument(
-        "input_item_identifier", help="Identifier for item for input to MWFR"
-    )
-    parser.add_argument(
-        "meta_workflow_identifier", help="MetaWorkflow identifier for input to MWFR"
-    )
-    parser.add_argument(
-        "-e", "--auth-env", default="default", help="Env name in ~/.cgap-keys.json"
-    )
-    parser.add_argument(
-        "--no-post", action="store_true", help="Do not POST the MWFR created"
-    )
-    parser.add_argument(
-        "--no-patch",
-        action="store_true",
-        help="Do not PATCH the input item with the MWFR created",
-    )
-    args = parser.parse_args()
-    auth_key = get_auth_key(args.auth_env)
-    post = not args.no_post
-    patch = not args.no_patch
-    create_meta_workflow_run(
-        args.input_item_identifier,
-        args.meta_workflow_identifier,
-        auth_key,
-        post=post,
-        patch=patch,
-    )
-
-
-if __name__ == "__main__":
-    main()
